refactor: move papyrus state reader
Move papyrus state reader into integration tests, since that is its only
use case.

** Strictly a move operation, no internal changes. **

commit-id:1bc1dc08
