+++ crates/gateway/Cargo.toml
@@ -16,19 +16,12 @@ axum.workspace = true
 blockifier.workspace = true
 cairo-lang-starknet-classes.workspace = true
 hyper.workspace = true
-indexmap.workspace = true
-papyrus_common.workspace = true
 papyrus_config.workspace = true
-papyrus_rpc.workspace = true
-papyrus_storage.workspace = true
 reqwest.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 starknet_api.workspace = true
-starknet_client.workspace = true
 starknet_mempool_types = { path = "../mempool_types", version = "0.0" }
-strum.workspace = true
-tempfile.workspace = true
 thiserror.workspace = true
 tokio.workspace = true
 validator.workspace = true
