+++ crates/tests-integration/src/integration_test_utils.rs
@@ -12,8 +12,9 @@ use starknet_gateway::config::{
 use starknet_gateway::errors::GatewayError;
 use starknet_gateway::gateway::Gateway;
 use starknet_gateway::starknet_api_test_utils::external_tx_to_json;
-use starknet_gateway::state_reader_test_utils::rpc_test_state_reader_factory;
 use starknet_mempool_types::communication::SharedMempoolClient;
+
+use crate::state_reader::rpc_test_state_reader_factory;
 
 pub async fn create_gateway(mempool_client: SharedMempoolClient) -> Gateway {
     let stateless_tx_validator_config = StatelessTransactionValidatorConfig {
