+++ crates/tests-integration/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod integration_test_setup;
 pub mod integration_test_utils;
+pub mod state_reader;
